requirements.txt
lightning==2.1.3
matplotlib==3.8.2
monai==1.3.0
monai_generative==0.2.2
numpy==1.26.3
pandas==2.2.0
Pillow==10.2.0
pytorch_lightning==2.1.3
PyYAML==6.0.1
PyYAML==6.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
